🎨 Moved admin toolbar helper inline with other ghost_head helpers
Removed the separate staff-frontend-tools service directory and inlined
the admin toolbar script assembly into ghost_head.js, matching the
pattern used by portal, search, and announcement-bar. Also added the
missing crossorigin="anonymous" attribute for CDN consistency.
