+++ ghost/core/core/frontend/helpers/ghost_head.js
@@ -14,7 +14,6 @@ const _ = require('lodash');
 const debug = require('@tryghost/debug')('ghost_head');
 const templateStyles = require('./tpl/styles');
 const {getFrontendAppConfig, getDataAttributes} = require('../utils/frontend-apps');
-const {getStaffFrontendHeadScripts} = require('../services/staff-frontend-tools/head-scripts');
 
 /**
  * @typedef {import('@tryghost/custom-fonts').FontSelection} FontSelection
@@ -135,6 +134,42 @@ function getAnnouncementBarHelper(data) {
     let helper = `<script defer src="${scriptUrl}" ${dataAttrs} crossorigin="anonymous"></script>`;
 
     return helper;
+}
+
+function getAdminToolbarHelper(dataRoot, siteTitle, excludeList) {
+    if (!dataRoot._locals?.staffFrontendToolsEnabled || excludeList.has('admin_toolbar')) {
+        return '';
+    }
+
+    const {scriptUrl} = getFrontendAppConfig('adminToolbar');
+    const context = dataRoot._locals?.context || dataRoot.context || [];
+    const entry = dataRoot.post || dataRoot.page;
+    const resourceId = entry?.id;
+    const resourceSlug = context.includes('tag') ? dataRoot.tag?.slug : '';
+    const isHome = context.includes('home');
+    let resourceType = '';
+
+    if (resourceId) {
+        resourceType = context.includes('page') || entry.type === 'page' ? 'page' : 'post';
+    } else if (resourceSlug) {
+        resourceType = 'tag';
+    }
+
+    const attrs = {
+        'ghost-admin-toolbar': escapeExpression(urlUtils.urlFor('admin', true)),
+        'site-title': escapeExpression(siteTitle || settingsCache.get('title') || 'Ghost'),
+        'resource-type': resourceType || undefined,
+        'resource-id': resourceId ? escapeExpression(resourceId) : undefined,
+        'resource-slug': resourceSlug ? escapeExpression(resourceSlug) : undefined,
+        'page-context': isHome ? 'home' : undefined,
+        'site-analytics-enabled': isHome && settingsCache.get('web_analytics_enabled') === true ? 'true' : undefined,
+        'activitypub-enabled': isHome && settingsCache.get('social_web_enabled') === true ? 'true' : undefined,
+        'members-enabled': isHome && settingsCache.get('members_enabled') === true ? 'true' : undefined,
+        'comments-enabled': resourceType === 'post' && settingsCache.get('comments_enabled') === 'off' ? 'false' : undefined
+    };
+    const dataAttrs = getDataAttributes(attrs);
+
+    return `<script defer src="${scriptUrl}" ${dataAttrs} crossorigin="anonymous"></script>`;
 }
 
 function getWebmentionDiscoveryLink() {
@@ -307,11 +342,10 @@ module.exports = async function ghost_head(options) { // eslint-disable-line cam
         if (!excludeList.has('announcement')) {
             head.push(getAnnouncementBarHelper(options.data));
         }
-        head.push(...getStaffFrontendHeadScripts({
-            dataRoot,
-            excludeList,
-            siteTitle: meta.site.title
-        }));
+        const adminToolbarHelper = getAdminToolbarHelper(dataRoot, meta.site.title, excludeList);
+        if (adminToolbarHelper) {
+            head.push(adminToolbarHelper);
+        }
         try {
             head.push(getWebmentionDiscoveryLink());
         } catch (err) {
